requirements.txt
# To install the packages listed in this file, use the command:
# pip install -r requirements.txt
numpy           ==  1.25.1
pillow          ==  11.0.0
pandas          ==  2.2.3
rdkit           ==  2024.3.6
pubchempy       ==  1.0.4
customtkinter   ==  5.2.2
pulp            ==  2.9.0
pywinstyles     ==  1.8